+++ Tests/userstories/src/test/java/com/skillstorm/definitions/deletedefinitions/DeleteItemSteps.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 
-import com.skillstorm.Selenium.ItemsPage;
+import com.skillstorm.pages.ItemsPage;
 
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
