+++ Tests/userstories/src/test/java/com/skillstorm/definitions/viewdefinitions/ViewAllWarehousesSteps.java
@@ -14,8 +14,6 @@ import com.skillstorm.pages.WarehousePage;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxOptions;
